telegram-bot/pyproject.toml
[project]
name = "telegram-rag-bot"
version = "0.1.0"
description = "A Telegram bot with RAG capabilities for document Q&A"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot[all]>=21.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.12.0",
    "markitdown[all]>=0.1.4",
    "nltk>=3.9.0",
    "tokenizers>=0.21.0",
    "transformers>=4.47.0",
    "openai>=1.58.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "black>=24.0.0",
    "mypy>=1.14.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true
